Make exception messages consistent
The same exceptions are thrown in several of the modules. To make sure
the messages are consistent they have been given their our module so it
is easy to share them.

The messages are returned by methods so the name of the argument for
which the exception is thrown can be added in the message. 