Use UTC time for time span calculation in calendar days
There was an error in the method to calculate a time span between two
dates and a result in calendar days. Because calculation were done in
local time problems occurred when the start and end date were not both
in DST or not in DST.

Example:
Getting the time span between March 28, 2019 (CET) and
April 1, 2019 (CEST) would return 3 days instead of 4. This because it
would subtract dates with a local time of midnight which is converted to
UTC time by the method getTime(). In reality the method calculated the
difference between March 27, at 23:00:00.000 UTC and March 31 at
22:00:00.000 UTC, which yields an unexpected result.